inputs: Add handling for shift
I tried to fit this in to the existing input infra. This seems a bit clunky to
me since AFAICT, SDL2 will handle most of this stuff pretty easily. I presume
that the current EN abstractions work really well for the general case, and
there is no point rocking the boat. It will get slightly worse with the next
patch when adding support for capslock.

Note that this patch actually makes more sense after capslock (next patch) - but
I wanted to do the simple one first in the series.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>