inputs: Add handling for capslock
Like the last patch this patch attempts to handle the 'shift' modifier within
the current EN input handling code. In this case, it handles capslock, which
unlike most other actions (except perhaps crouch), has a state.

I believe there is probably a better way to do this with SDL, and perhaps even
the EN input code can utilize it, but I couldn't figure out how to do it within
the existing EN framework.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>